# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Invoices API",
        "description": "API de controle financeiro com fluxo de caixa e separação por carteira. Permite lançar contas a pagar e a receber de forma única, parcelada ou mensal.",
        "contact": {
            "name": "My Invoices API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realizar login",
                "description": "Autentica um usuário e retorna um token de acesso",
                "operationId": "ec9c379d2a959de66404708dd4a9322d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 4,
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login realizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "token": "1|abcdef1234567890",
                                    "message": "Login realizado com sucesso"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário ou senha inválidos"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "token": null,
                                    "message": "Usuário ou senha inválidos"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário ou senha inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar novo usuário",
                "description": "Cria uma nova conta de usuário. Um email de ativação será enviado.",
                "operationId": "f367ea894b2ac44a9607f8e0a9a03537",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 4,
                                        "example": "senha123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "João Silva",
                                "email": "usuario@example.com",
                                "password": "senha123",
                                "password_confirmation": "senha123"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "usuario@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário cadastrado com sucesso, acesse seu e-mail e clique no link de ativação"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "user": {
                                        "name": "João Silva",
                                        "email": "usuario@example.com"
                                    },
                                    "message": "Usuário cadastrado com sucesso, acesse seu e-mail e clique no link de ativação"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao registrar usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email já cadastrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dados inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activate": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Ativar conta de usuário",
                "description": "Ativa a conta do usuário usando o token recebido por email",
                "operationId": "4623874454f119c4a5df5f2a71d07caa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "abc123def456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário ativado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário ativado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token inválido ou expirado"
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    }
                }
            }
        },
        "/forgot-password": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Solicitar recuperação de senha",
                "description": "Envia um email com token para redefinição de senha",
                "operationId": "9e273ee835bbbe0d6ad487331f877940",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email de recuperação enviado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "E-mail de redefinição de senha enviado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email não encontrado"
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    }
                }
            }
        },
...
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realizar logout",
                "description": "Revoga todos os tokens de autenticação do usuário",
                "operationId": "cc14526f9abd470d63047b1c05aef0c3",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}